Use relative path so sourcemap hash consistently
esbuild supports outputting files with content-addressable names,
including sourcemaps

This path being absolute means the output hash depends on the absolute
path the build happened in, which makes hashes change every time when
building in CI environments like Jenkins where the build path depends on
the job name, PR name, and sequential build number.

One drawback of this change is the path now doesn't map to a file in the
filesystem. I tried to find a way to guarantee it would, but failed.